Refactor class names and package structure
