--- core/src/com/simanglam/PokemonWarrior.java
@@ -1,5 +1,0 @@
-package com.simanglam;
-
-public class PokemonWarrior {
-    
-}
