+++ core/src/com/simanglam/fighting/Pokemon.java
@@ -1,4 +1,6 @@
-package com.simanglam;
+package com.simanglam.fighting;
+
+import com.simanglam.fighting.bosswar.PokemonWarrior;
 
 public interface Pokemon {
     public String getName();
