+++ core/src/com/simanglam/fighting/bosswar/BossWarActor.java
@@ -1,0 +1,7 @@
+package com.simanglam.fighting.bosswar;
+
+import com.badlogic.gdx.scenes.scene2d.Actor;
+
+public abstract class BossWarActor extends Actor{
+    
+}
