+++ core/src/com/simanglam/fighting/bosswar/PokemonWarrior.java
@@ -1,0 +1,7 @@
+package com.simanglam.fighting.bosswar;
+
+import com.badlogic.gdx.scenes.scene2d.Actor;
+
+public class PokemonWarrior extends Actor{
+    
+}
