+++ core/src/com/simanglam/map/World.java
@@ -1,4 +1,4 @@
-package com.simanglam;
+package com.simanglam.map;
 
 import com.badlogic.gdx.graphics.OrthographicCamera;
 import com.badlogic.gdx.graphics.g2d.SpriteBatch;
@@ -13,6 +13,7 @@ import com.badlogic.gdx.math.Intersector;
 import com.badlogic.gdx.math.Rectangle;
 import com.badlogic.gdx.utils.viewport.FitViewport;
 import com.badlogic.gdx.utils.viewport.Viewport;
+import com.simanglam.util.Const;
 
 public class World {
     public TiledMap tiledMap;
