+++ core/src/com/simanglam/map/ui/Dialog.java
@@ -1,4 +1,4 @@
-package com.simanglam;
+package com.simanglam.map.ui;
 
 import java.util.ArrayList;
