+++ core/src/com/simanglam/scenes/InfoScreen.java
@@ -10,8 +10,8 @@ import com.badlogic.gdx.scenes.scene2d.ui.Label;
 import com.badlogic.gdx.scenes.scene2d.ui.Skin;
 import com.badlogic.gdx.scenes.scene2d.ui.Window;
 import com.badlogic.gdx.utils.viewport.ExtendViewport;
-import com.simanglam.Const;
 import com.simanglam.Main;
+import com.simanglam.util.Const;
 
 public class InfoScreen extends InputAdapter implements Screen {
     Main game;
