+++ core/src/com/simanglam/util/InventoryItem.java
@@ -1,4 +1,4 @@
-package com.simanglam;
+package com.simanglam.util;
 
 public interface InventoryItem {
